BUGFIX: make moving of nodes more robust
Moving of nodes did not work when the *rendered parent node* in DOM was not
the parent node in the node tree.

In this case, the following happened:

- the rendered parent node was transmitted as "parentDomAddress.contextPath"
  and made available in AbstractChange as getParentNode()
- in AbstractMove::canApply(), the "getParentNode()" method was called, and
  taken into account for constraint checks -- so if the rendered parent node
  was a different node than the actual parent (e.g. the grandparent), the
  constraints might not have matched and the move could not succeed.

To fix this, I changed the following:

- removed AbstractStructuralChange::getParentNode as this is not really the parent node, but the closest rendered node.
- pushed canApply() down; not relying on getParentNode() anymore
- for the Copy/Move INTO cases, add a specific property "parentContextPath" to the
  change; which contains the parent node without further processing.
- we ensure the renderedParentNode === parentNode (==is the ContentCollection),
  otherwise out of band rendering can not work (and we need to reload the full page
  instead).
