+++ Classes/Neos/Neos/Ui/Domain/Model/Changes/CopyAfter.php
@@ -13,6 +13,20 @@ namespace Neos\Neos\Ui\Domain\Model\Changes;
 
 class CopyAfter extends AbstractCopy
 {
+
+    /**
+     * "Subject" is the to-be-copied node; the "sibling" node is the node after which the "Subject" should be copied.
+     *
+     * @return boolean
+     */
+    public function canApply()
+    {
+        $nodeType = $this->getSubject()->getNodeType();
+
+        return $this->getSiblingNode()->getParent()->isNodeTypeAllowedAsChildNode($nodeType);
+    }
+
+
     public function getMode()
     {
         return 'after';
@@ -26,7 +40,7 @@ class CopyAfter extends AbstractCopy
     public function apply()
     {
         if ($this->canApply()) {
-            $nodeName = $this->generateUniqueNodeName($this->getParentNode());
+            $nodeName = $this->generateUniqueNodeName($this->getSiblingNode()->getParent());
             $node = $this->getSubject()->copyAfter($this->getSiblingNode(), $nodeName);
             $this->finish($node);
         }
