+++ Classes/Neos/Neos/Ui/Domain/Model/Changes/CopyInto.php
@@ -1,4 +1,5 @@
 <?php
+
 namespace Neos\Neos\Ui\Domain\Model\Changes;
 
 /*
@@ -11,8 +12,57 @@ namespace Neos\Neos\Ui\Domain\Model\Changes;
  * source code.
  */
 
+use Neos\ContentRepository\Domain\Model\NodeInterface;
+
 class CopyInto extends AbstractCopy
 {
+
+
+    /**
+     * @var string
+     */
+    protected $parentContextPath;
+
+    /**
+     * @var NodeInterface
+     */
+    protected $cachedParentNode;
+
+    /**
+     * @param string $parentContextPath
+     */
+    public function setParentContextPath($parentContextPath)
+    {
+        $this->parentContextPath = $parentContextPath;
+    }
+
+    /**
+     * @return NodeInterface
+     */
+    public function getParentNode()
+    {
+        if ($this->cachedParentNode === null) {
+            $this->cachedParentNode = $this->nodeService->getNodeFromContextPath(
+                $this->parentContextPath
+            );
+        }
+
+        return $this->cachedParentNode;
+    }
+
+    /**
+     * "Subject" is the to-be-copied node; the "parent" node is the new parent
+     *
+     * @return boolean
+     */
+    public function canApply()
+    {
+        $nodeType = $this->getSubject()->getNodeType();
+
+        return $this->getParentNode()->isNodeTypeAllowedAsChildNode($nodeType);
+    }
+
+
     public function getMode()
     {
         return 'into';
