+++ Classes/Neos/Neos/Ui/Domain/Model/Changes/MoveAfter.php
@@ -15,6 +15,19 @@ use Neos\Neos\Ui\Domain\Model\Feedback\Operations\UpdateNodeInfo;
 
 class MoveAfter extends AbstractMove
 {
+    /**
+     * "Subject" is the to-be-moved node; the "sibling" node is the node after which the "Subject" should be copied.
+     *
+     * @return boolean
+     */
+    public function canApply()
+    {
+        $nodeType = $this->getSubject()->getNodeType();
+
+        return $this->getSiblingNode()->getParent()->isNodeTypeAllowedAsChildNode($nodeType);
+    }
+
+
     public function getMode()
     {
         return 'after';
