+++ packages/neos-ui/src/Sagas/CR/NodeOperations/helpers.js
@@ -49,6 +49,7 @@ export const calculateDomAddressesFromMode = (mode, contextPath, fusionPath) => 
             const element = findNodeInGuestFrame(contextPath, fusionPath);
 
             return {
+                parentContextPath: contextPath,
                 parentDomAddress: {
                     contextPath: element ? element.getAttribute('data-__neos-node-contextpath') : contextPath,
                     fusionPath: element ? element.getAttribute('data-__neos-fusion-path') : fusionPath
